DuckDB to use the Scan API
